--- README.md
@@ -1,65 +1,0 @@
-# Sendbird Desk for JavaScript sample
-
-![Platform](https://img.shields.io/badge/platform-JAVASCRIPT-orange.svg)
-![Languages](https://img.shields.io/badge/language-JAVASCRIPT-orange.svg)
-
-## Introduction
-
-Built with Sendbird Chat platform, Sendbird Desk is a live chat customer support that offers customer satisfaction through enhanced engagement. Through its integration, Desk SDK for JavaScript enables you to easily customize your ticketing support system with a UI theme, thereby elevating your overall customers’ experience. For example, you can modify the inbox - a management tool and storage unit for the agents’ and tickets’ conversations - to fit within your color scheme and layout.
-
-### More about Sendbird Desk for JavaScript
-
-Find out more about Sendbird Desk for JavaScript on [Desk for JavaScript doc](https://sendbird.com/docs/desk/v1/javascript/getting-started/about-desk-sdk). If you need any help in resolving any issues or have questions, visit [our community](https://community.sendbird.com).
-
-<br />
-
-## Before getting started
-
-This section shows you the prerequisites you need for testing Sendbird Desk for JavaScript sample app.
-
-### Requirements
-
-- NodeJS 10.13.x+ along with NPM
-
-### Browser support
-
-- Modern browsers supporting ES6+ (Chrome, FireFox, Edge, Safari, etc)
-- IE 11+
-- Mobile browsers (Android/iOS)
-
-### Try the sample app using your data 
-
-If you would like to customize the sample app for your usage, you can replace the default sample app ID with your ID - which you can obtain by [creating your Sendbird application from the dashboard](https://sendbird.com/docs/chat/v3/ios/getting-started/install-chat-sdk#2-step-1-create-a-sendbird-application-from-your-dashboard).
-
-> Note: After creating the Sendbird application, please be sure to contact [sales](https://get.sendbird.com/talk-to-sales.html) to enable the **Desk** menu onto the dashboard. Currently, Sendbird Desk is available only for **free-trial** or **Enterprise** plans.
-
-Following the previous instructions will allow you to experience the sample app with your data from the Sendbird application.
-
-<br />
-
-## Getting started
-
-This section explains the steps you need to take before testing the sample app.
-
-### Use npm
-
-Using `npm`, you can install and build the sample as follows: 
-
-```bash
-npm install
-npm run build
-```
-
-### Run demo
-
-Access local server by typing `http://127.0.0.1:8888` on your web browser after executing the command as follows: 
-
-```bash
-npm run demo
-```
-
-<br />
-
-## For further reference
-
-Please visit the following link to learn more about Desk SDK for JavaScript: https://github.com/sendbird/SendBird-Desk-SDK-JavaScript
