+++ index.html
@@ -1,0 +1,24 @@
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <link rel="icon" type="image/svg+xml" href="/img/sendbird-logo.png" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link href="./src/widget/sass/widget.scss" rel="stylesheet" />
+    <title>Sendbird desk</title>
+  </head>
+  <body>
+    <div class='guide'>
+      <div class='logo'><img src='img/sendbird-logo.png' /></div>
+      <div class='title'><span class='bold'>Desk Widget</span> Sample Page</div>
+      <div class='desc'>This is the Desk Widget Sample page.<br />To try our Desk Widget, please click the button below.</div>
+      <img src='img/guide-arrow.png' class='arrow' />
+    </div>
+    <div class='guide-balloon'>
+        <div class='emoji'>😀</div>
+        <div class='desc'>To try Desk Widget, please click the icon.</div>
+    </div>
+    <div id='sb-desk' class='-sbd-widget'></div>
+    <script type="module" src="./src/widget/js/login.js"></script>
+  </body>
+</html>
