+++ src/widget/js/component/notification.js
