+++ src/widget/js/component/spinner.js
