+++ src/widget/js/domparser.js
@@ -1,8 +1,10 @@
 import { simplify } from './simplify.js';
+import DOMPurify from 'dompurify';
 
 export const parseDom = es => {
   var $ = document.createElement('div');
-  $.innerHTML = es.trim();
+  let clean = DOMPurify.sanitize(es);
+  $.innerHTML = clean.trim();
   $ = simplify($.firstChild);
   if ($.parentNode) $.parentNode.removeChild($);
   return $;
