+++ src/widget/js/globalStore.js
@@ -1,0 +1,11 @@
+const store = {
+  sb: null,
+}
+
+export function setSb(sb) {
+  store.sb = sb
+}
+
+export function getSb() {
+  return store.sb
+}
