+++ src/widget/sass/mixin.scss
@@ -1,5 +1,4 @@
 /*$ic-prefix: 'https://dxstmhyqfqr1o.cloudfront.net/desk/widget/';*/
-// $ic-prefix: '/desk/widget/img/'; // Use this prefix when you deploy to the server
 $ic-prefix: '/img/';
 
 @mixin hover {
