+++ src/widget/sass/widget.scss
@@ -676,76 +676,6 @@ body {
                   }
                 }
               }
-              .chatbot-card {
-                position: relative;
-                display: inline-flex;
-                flex-direction: column;
-
-                .faq-card {
-                  position: relative;
-                  display: flex;
-                  width: 264px;
-                  min-width: 264px;
-                  height: 82px;
-                  border: solid 1px #ccc;
-                  border-radius: 8px;
-                  background-color: #ffffff;
-                  margin-top: 4px;
-                  .faq-card__profile-image {
-                    position: absolute;
-                    width: 56px;
-                    height: 56px;
-                    top: 12px;
-                    left: 12px;
-                  }
-                  .faq-card__title-question {
-                    position: absolute;
-                    top: 12px;
-                    left: 80px;
-                    width: 164px;
-                    height: 20px;
-                    // font
-                    font-size: 14px;
-                    font-weight: 500;
-                    font-stretch: normal;
-                    font-style: normal;
-                    line-height: 1.43;
-                    letter-spacing: -0.1px;
-                    text-align: left;
-                    color: #0d0d0d;
-                    overflow: hidden;
-                    text-overflow: ellipsis;
-                    white-space: nowrap;
-                  }
-                  .faq-card__description {
-                    position: absolute;
-                    top: 36px;
-                    left: 80px;
-                    width: 164px;
-                    height: 32px;
-                    // reset margin from p tag
-                    margin-block-start: 0px;
-                    margin-block-end: 0px;
-                    // font
-                    font-size: 12px;
-                    font-weight: normal;
-                    font-stretch: normal;
-                    font-style: normal;
-                    line-height: 1.33;
-                    letter-spacing: normal;
-                    text-align: left;
-                    color: #858585;
-                    overflow: hidden;
-                    text-overflow: ellipsis;
-                    // multiple line ellipsis
-                    display: -webkit-box;
-                    max-width: 200px;
-                    -webkit-line-clamp: 4;
-                    -webkit-box-orient: vertical;
-                    overflow: hidden;
-                  }
-                }
-              }
               .created-at {
                 font-size: 12px;
                 color: #999;
